Need to convert floats to doubles before handing them to printf.
git-svn-id: https://smlnj-gforge.cs.uchicago.edu/svn/manticore/trunk@1955 ac4fbcad-d21d-0410-b760-c7b1d3380187

+++ src/lib/parallel-rt/misc/basis.c
@@ -37,7 +37,7 @@ Value_t M_LongToString (int64_t n)
 Value_t M_FloatToString (float f)
 {
     char buf[64];
-    snprintf(buf, sizeof(buf), "%f", f);
+    snprintf(buf, sizeof(buf), "%f", (double)f);
     return AllocString (VProcSelf(), buf);
 }
 
@@ -207,7 +207,7 @@ void M_PrintInt (int32_t n)
 
 void M_PrintFloat (float f)
 {
-    Say ("%f\n",f);
+  Say ("%f\n",(double)f);
 }
 
 int M_ReadInt ()
